add tty to digital ocean config. update img run to use tty, and ssh port forward

+++ etc/nixos/digital-ocean/img_run.sh
@@ -1,5 +1,16 @@
 new_img=/tmp/gen_image.qcow2
-rm -rf $new_img
-cp ./result/nixos-image-qcow2* $new_img
-chmod 666 $new_img
-qemu-system-x86_64 -enable-kvm -m 2048 -nic user,model=virtio -drive file=$new_img,media=disk,if=virtio
+
+if [ -e ./result ]; then
+  rm -rf $new_img
+  cp ./result/nixos-image-qcow2* $new_img
+  chmod 666 $new_img
+fi
+
+qemu-system-x86_64 \
+-enable-kvm \
+-nographic \
+-serial mon:stdio \
+-m 2048 \
+-nic user,model=virtio,hostfwd=tcp::8022-:22 \
+-drive file=$new_img,media=disk,if=virtio
+
